+++ example/lib/module/refresh_page.dart
@@ -2,7 +2,6 @@ import 'package:app/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 
-
 class _Item extends StatelessWidget {
   const _Item(this.index, this.color);